apiTab21

+++ src/tabs/use.Tab.hook.tsx
@@ -1,6 +1,4 @@
 import React, { useCallback, useEffect, useState } from 'react';
-import { ActiveInputService } from '../services/ActiveInputService';
-import { Hints } from '../custom/variables';
 
 
 export interface TabItems {
@@ -9,11 +7,6 @@ export interface TabItems {
 	items: AddedItem[];
 }
 
-
-enum ItemTypes {
-	weights = 'weighed',
-	piece = 'pieced',
-}
 
 interface Defaults {
 	tara: number;
@@ -47,107 +40,32 @@ export interface AddedItem extends Item {
 	cost: number;
 }
 
-export interface ArgAddItemFunc {
-	item: Item,
-	calcValue?: number
-}
-
-interface TabId {
+export interface TabId {
 	id: number;
 }
 
-export function useTabs(
-	setHint: ( str: Hints, likeError?: boolean ) => void,
-	scaleService: any,
-	calcValue: number,
-): [
-	TabItems[],
+export function useTabs(): [
+	TabId[],
 	number,
-		AddedItem | null,
+	() => void,
+	() => void,
 	React.Dispatch<React.SetStateAction<number>>,
-	React.Dispatch<React.SetStateAction<AddedItem | null>>,
-	( { item, calcValue }: ArgAddItemFunc ) => boolean,
-	() => void,
-	() => void,
-	() => void,
-	( tara: number ) => void, // setTara
-	() => void, // print
-	() => number // getTara
 ] {
 
 	const _apiBase = `http://10.13.16.80:4445`;
 
-
-	const [ tabs, setTabs ] = useState<Array<TabId>>( [] )
+	const [ tabs, setTabs ] = useState<Array<TabId>>( [ { id: 1 } ] )
+	const [ activeTab, setActiveTab ] = useState<number>( tabs[0].id )
 
 	useEffect( () => {
 		fetch( `${ _apiBase }/tab/list` )
 		.then( ( res ) => res.json() )
-		.then( value => setTabs( value ) )
+		.then( ( value ) => {
+			setTabs( value )
+			setActiveTab( tabs[tabs.length - 1].id )
+		} )
 	}, [] )
 
-	const [ tabItems, setTabItems ] = useState<TabItems[]>( [ { tabNumber: 1, tara: 0, items: [] } ] as TabItems[] );
-	const [ activeTab, setActiveTab ] = useState<number>( 0 );
-	const [ activeItem, setActiveItem ] = useState<AddedItem | null>( null );
-
-	const setTara = useCallback( ( tara ) => {
-		tabItems[activeTab].tara = tara;
-		setTabItems( [ ...tabItems ] );
-	}, [ tabItems, activeTab ] );
-	const getTara = useCallback( () => {
-		return (tabItems[activeTab].tara) / 1000
-	}, [ tabItems, activeTab ] );
-	const addItem = useCallback(
-		( { item, calcValue }: ArgAddItemFunc ) => {
-			if ( scaleService.checkStable() ) {
-				const addedItem = { ...item } as AddedItem;
-				scaleService.setTitle( item.texts.full_title );
-				scaleService.setPrice( item.price );
-				const weightScale = scaleService.getItemWeight();
-				if ( weightScale >= (40 / 1000) ) {
-					switch ( item.type ) {
-						case ItemTypes.piece:
-							if ( calcValue ) {
-								addedItem.amount = calcValue;
-								addedItem.cost = addedItem.amount * item.price;
-							} else {
-								setHint( Hints.PickItemsQty );
-								return false;
-							}
-							break;
-						case ItemTypes.weights:
-							addedItem.amount = weightScale;
-							addedItem.cost = scaleService.getItemCost();
-							break;
-						default:
-							setHint( Hints.IncorrectItemType, true );
-							return false;
-					}
-				} else {
-					setHint( Hints.MinWeight, true );
-					return false;
-				}
-				tabItems[activeTab].items.push( addedItem );
-				setTabItems( [ ...tabItems ] );
-				ActiveInputService.clear();
-				return true;
-			}
-			return false;
-		}
-		,
-		[ tabItems, activeTab ],
-	);
-	const deleteItem = useCallback( () => {
-		tabItems[activeTab].items = tabItems[activeTab].items.filter( ( item ) => item !== activeItem );
-		setTabItems( [ ...tabItems ] );
-		setActiveItem( null );
-	}, [ tabItems, activeItem ] );
-
-	const print = useCallback( () => {
-		console.log( '-------------------------' )
-		console.log( 'print', tabItems[activeTab].items )
-		console.log( '-------------------------' )
-	}, [ activeTab, tabItems ] );
 
 	const createTab = useCallback( () => {
 		fetch( `${ _apiBase }/create-tab`, {
@@ -156,43 +74,41 @@ export function useTabs(
 		.then( ( res ) => res.json() )
 		.then( ( tabId: any ) => setTabs( prevState => {
 			if ( !tabId.id ) return prevState
-			return [ ...prevState, tabId.id ]
+			const newState = [ ...prevState, tabId.id ]
+			setActiveTab( newState[newState.length - 1].id )
+			return newState
 		} ) )
-	}, [] )
+	}, [ activeTab, tabs ] )
 
 	const deleteTab = useCallback( () => {
-		const id = 2
+		// if ( tabs.length <= 1 && tabs[tabs.length - 1].id === 1 ) return
+		if ( tabs.length <= 1 && tabs[tabs.length - 1].id <= 1 ) return
 		fetch( `${ _apiBase }/delete-tab`, {
 			method: 'DELETE',
-			body: JSON.stringify( { "id": `${ id }` } ),
+			body: JSON.stringify( { "id": `${ activeTab }` } ),
 			headers: { 'Content-type': 'application/json' }
 		} )
 		.then( ( res ) => res.json() )
 		.then( ( res: any ) => setTabs( ( prevState ) => {
-			console.log( res )
+			console.log( 'deleteTab', res )
 			if ( !res.affected ) {
 				return prevState
 			}
-			return prevState.filter( ( num ) => num.id !== id )
+			const newState = prevState.filter( ( num ) => num.id !== activeTab )
+			setActiveTab( newState[0].id )
+			return newState
 		} ) )
-	}, [] )
+	}, [ activeTab, tabs ] )
 
 	useEffect( () => {
-		console.log( tabs )
-	}, [ tabs ] )
+		console.log( 'tabs', tabs )
+	}, [ tabs, activeTab ] )
 
 	return [
-		tabItems,
+		tabs,
 		activeTab,
-		activeItem,
-		setActiveTab,
-		setActiveItem,
-		addItem,
-		deleteItem,
 		createTab,
 		deleteTab,
-		setTara,
-		print,
-		getTara
+		setActiveTab
 	];
 }
